style: move tx function
